requirements.txt
Django==4.2.8
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2==2.9.9
python-decouple==3.8
django-filter==23.4